# src/docs/swagger.yaml
basePath: /
definitions:
...
  httputil.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
...
  model.Favorite:
    properties:
      favoritedAt:
        type: string
      id:
        type: integer
      uid:
        type: string
      vendorId:
        type: string
    type: object
...
  model.ItemWithdrawal:
    properties:
      fulfilledAt:
        type: string
      id:
        type: integer
      userItemId:
        type: integer
      withdrawnAt:
        type: string
    type: object
...
  model.PageUserFavorite:
    properties:
      favoriteId:
        type: integer
      favoritedAt:
        type: string
      favoritesAmount:
        type: integer
      uid:
        type: string
      vendorActive:
        type: boolean
      vendorBannerUrl:
        type: string
      vendorBio:
        type: string
      vendorBirthday:
        type: string
      vendorCreatedAt:
        type: string
      vendorDisplayName:
        type: string
      vendorFavoritesAmount:
        type: integer
      vendorFirstName:
        type: string
      vendorId:
        type: string
      vendorLastName:
        type: string
      vendorMainUrl:
        type: string
      vendorPackAmount:
        type: integer
      vendorThumbUrl:
        type: string
      vendorUpdatedAt:
        type: string
      vendorUsername:
        type: string
    type: object
  model.PageUserItem:
    properties:
      acquiredAt:
        type: string
      active:
        type: boolean
      categories:
        items:
          type: string
        type: array
      contentExpiredTime:
        type: string
      contentMainUrl:
        type: string
      contentThumbUrl:
        type: string
      contentType:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      description:
        type: string
      expiredAt:
        type: string
      imageUrl:
        type: string
      itemId:
        type: integer
      name:
        type: string
      notify:
        type: boolean
      ownerId:
        type: string
      ranking:
        type: integer
      rarity:
        type: string
      rawCategories:
        type: string
      totalItemsAmount:
        type: integer
      updatedAt:
        type: string
      userItemId:
        type: integer
      value:
        type: number
      vendorActive:
        type: boolean
      vendorBannerUrl:
        type: string
      vendorBio:
        type: string
      vendorDisplayName:
        type: string
      vendorFavoritesAmount:
        type: integer
      vendorFirstName:
        type: string
      vendorId:
        type: string
      vendorImageUrl:
        type: string
      vendorLastName:
        type: string
      vendorMainUrl:
        type: string
      vendorThumbUrl:
        type: string
      vendorUsername:
        type: string
    type: object
  model.PageUserPack:
    properties:
      categories:
        items:
          type: string
        type: array
      contentMainUrl:
        type: string
      contentThumbUrl:
        type: string
      createdAt:
        type: string
      description:
        type: string
      imageUrl:
        type: string
      itemQty:
        type: integer
      latestPurchasedAt:
        type: string
      packAmount:
        type: integer
      packConfigId:
        type: integer
      packIds:
        items:
          type: integer
        type: array
      qty:
        type: integer
      rawCategories:
        type: string
      rawPackIds:
        type: string
      title:
        type: string
      tokenAmount:
        type: number
      totalPacksAmount:
        type: integer
      vendorActive:
        type: boolean
      vendorBannerUrl:
        type: string
      vendorBio:
        type: string
      vendorDisplayName:
        type: string
      vendorFavoritesAmount:
        type: integer
      vendorFirstName:
        type: string
      vendorId:
        type: string
      vendorImageUrl:
        type: string
      vendorLastName:
        type: string
      vendorMainUrl:
        type: string
      vendorThumbUrl:
        type: string
      vendorUsername:
        type: string
    type: object
...
  model.User:
    properties:
      active:
        type: boolean
      bannerImgVersion:
        type: integer
      bannerUrl:
        type: string
      bio:
        type: string
      birthday:
        type: string
      contentMainUrl:
        type: string
      contentThumbUrl:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      displayName:
        type: string
      email:
        type: string
      firstName:
        type: string
      imageUrl:
        type: string
      isVendor:
        type: boolean
      lastName:
        type: string
      lastUsernameChangeAt:
        type: string
      profileImgVersion:
        type: integer
      uid:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
  model.UserFavoritePage:
    properties:
      favoriteAmount:
        type: integer
      nextPage:
        type: integer
      page:
        items:
          $ref: '#/definitions/model.PageUserFavorite'
        type: array
      pageSize:
        type: integer
    type: object
  model.UserItemPage:
    properties:
      earliestExpired:
        type: integer
      nextPage:
        type: integer
      page:
        items:
          $ref: '#/definitions/model.PageUserItem'
        type: array
      pageSize:
        type: integer
      userItemAmount:
        type: integer
    type: object
  model.UserPackPage:
    properties:
      nextPage:
        type: integer
      page:
        items:
          $ref: '#/definitions/model.PageUserPack'
        type: array
      pageSize:
        type: integer
      userPackAmount:
        type: integer
    type: object
...
host: apidev.xopacks.com
info:
  contact: {}
  description: The official XOPAcks API
  title: XOPacks API
  version: "1.0"
paths:
...
  /user:
    delete:
      consumes:
      - application/json
      description: Deactivates a user from the DB
      parameters:
      - description: uid
        in: query
        name: uid
        required: true
        type: string
      - description: username
        in: query
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Delete a user
      tags:
      - User
    patch:
      consumes:
      - application/json
      description: Updates a user record in DB given a patch map
      parameters:
      - description: uid
        in: query
        name: uid
        required: true
        type: string
      - description: username
        in: query
        name: username
        required: true
        type: string
      - description: user patch map
        in: body
        name: userPatchMap
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Patch a user
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Save user record in DB
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Create a user
      tags:
      - User
  /user/{uid}:
    get:
      consumes:
      - application/json
      description: Save a user record in DB
      parameters:
      - description: uid
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Get a user
      tags:
      - User
  /user/favorite:
    delete:
      consumes:
      - application/json
      description: Add a vendor favorite record to a user
      parameters:
      - description: uid
        in: query
        name: uid
        required: true
        type: string
      - description: vendor uid
        in: query
        name: vendorId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Favorite'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Add a user favorite
      tags:
      - User
    get:
      consumes:
      - application/json
      description: Get a favorite between user and vendor
      parameters:
      - description: uid
        in: query
        name: uid
        required: true
        type: string
      - description: vendor uid
        in: query
        name: vendorId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Favorite'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Get a favorite
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Add a vendor favorite record to a user
      parameters:
      - description: uid
        in: query
        name: uid
        required: true
        type: string
      - description: vendor uid
        in: query
        name: vendorId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Favorite'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Add a user favorite
      tags:
      - User
  /user/favorites/{uid}:
    get:
      consumes:
      - application/json
      description: Get user favorites vendor page
      parameters:
      - description: vendor uid
        in: path
        name: uid
        required: true
        type: string
      - description: page number
        in: query
        name: pageNum
        required: true
        type: integer
      - description: search string
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserFavoritePage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Get user favorite page
      tags:
      - User
  /user/item/withdrawal:
    post:
      consumes:
      - application/json
      description: Allows a user to withdrawal an item from their collection which
        is externally fulfilled
      parameters:
      - description: uid
        in: query
        name: uid
        required: true
        type: string
      - description: authorized uid
        in: query
        name: authorizedUid
        required: true
        type: string
      - description: user item id
        in: query
        name: userItemId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.ItemWithdrawal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Withdrawal a user item
      tags:
      - User
  /user/items/{uid}:
    get:
      consumes:
      - application/json
      description: Get a list of items that a user has in their collections
      parameters:
      - description: vendor uid
        in: path
        name: uid
        required: true
        type: string
      - description: page number
        in: query
        name: pageNum
        required: true
        type: integer
      - description: sort by
        in: query
        name: sortBy
        type: string
      - description: sort direction
        in: query
        name: sortDir
        type: string
      - description: category filter
        in: query
        name: filterOn
        type: string
      - description: search string
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserItemPage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Get user collection item page
      tags:
      - User
  /user/packs/{uid}:
    get:
      consumes:
      - application/json
      description: Get a list of packs that a user has in their collections
      parameters:
      - description: vendor uid
        in: path
        name: uid
        required: true
        type: string
      - description: page number
        in: query
        name: pageNum
        required: true
        type: integer
      - description: sort by
        in: query
        name: sortBy
        type: string
      - description: sort direction
        in: query
        name: sortDir
        type: string
      - description: category filter
        in: query
        name: filterOn
        type: string
      - description: search string
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserPackPage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Get user collection pack page
      tags:
      - User
  /user/username/:username:
    get:
      consumes:
      - application/json
      description: Retrieve a user given the username
      parameters:
      - description: username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Get a user by username
      tags:
      - User
...
swagger: "2.0"